# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance media conversion API delivering WhatsApp-ready audio and images.\n\n**Author:** Guilherme Jansen · [GitHub](https://github.com/guilhermejansen)\n**Contato direto:** suporte@setupautomatizado.com.br\n**Coleção Postman pronta:** [Download](./swagger/postman.json)\n\nUse os exemplos documentados em cada endpoint para testar rapidamente conversões de áudio, imagem e operações S3.",
        "title": "WhatsApp Media Converter API",
        "contact": {
            "name": "Guilherme Jansen",
            "url": "https://github.com/guilhermejansen",
            "email": "suporte@setupautomatizado.com.br"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/upload/s3": {
            "post": {
                "description": "Accepts large media as multipart form-data and dispatches asynchronous upload jobs.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "S3"
                ],
                "summary": "Start multipart upload to S3-compatible storage",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Binary file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JSON encoded upload options",
                        "name": "options",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResponse"
                        }
                    }
                }
            }
        },
        "/upload/s3/base64": {
            "post": {
                "description": "Accepts pre-encoded data URIs and dispatches asynchronous upload jobs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "S3"
                ],
                "summary": "Start base64 upload to S3-compatible storage",
                "parameters": [
                    {
                        "description": "Base64 upload request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3Base64UploadRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResponse"
                        }
                    }
                }
            }
        },
...
        "/upload/s3/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "S3"
                ],
                "summary": "List recent upload jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by upload status (pending|uploading|completed|failed|cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadListResponse"
                        }
                    }
                }
            }
        },
        "/upload/s3/object/{key}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "S3"
                ],
                "summary": "Retrieve object metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Object key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_providers.ObjectInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "S3"
                ],
                "summary": "Delete object from storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Object key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/upload/s3/status/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "S3"
                ],
                "summary": "Retrieve asynchronous upload status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Upload identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "S3"
                ],
                "summary": "Cancel an in-flight upload job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Upload identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "whats-convert-api_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Missing 'data' field"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "whats-convert-api_internal_models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "whats-convert-api_internal_models.S3Base64UploadRequest": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string",
                    "example": "audio/ogg"
                },
                "data": {
                    "type": "string",
                    "example": "data:audio/ogg;base64,T2dnUwACAAAAAAAAAAB"
                },
                "expires_days": {
                    "type": "integer",
                    "example": 3
                },
                "filename": {
                    "type": "string",
                    "example": "voice-note.opus"
                },
                "key": {
                    "type": "string",
                    "example": "uploads/audio/voice-note.opus"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "public": {
                    "type": "boolean",
                    "example": false
                },
                "storage_class": {
                    "type": "string",
                    "example": "STANDARD"
                }
            }
        },
...
        "whats-convert-api_internal_models.S3UploadListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 1
                },
                "uploads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadStatusResponse"
                    }
                }
            }
        },
...
        "whats-convert-api_internal_models.S3UploadResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to start upload"
                },
                "message": {
                    "type": "string",
                    "example": "Upload started successfully"
                },
                "result": {
                    "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResult"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "upload_id": {
                    "type": "string",
                    "example": "3f99d60f-bd8d-49e6-9ecf-2fbc9e4adffe"
                }
            }
        },
        "whats-convert-api_internal_models.S3UploadResult": {
            "type": "object",
            "properties": {
                "etag": {
                    "type": "string",
                    "example": "\"9b2cf535f27731c974343645a3985328\""
                },
                "expires_at": {
                    "type": "string",
                    "example": "2024-04-01T12:00:00Z"
                },
                "key": {
                    "type": "string",
                    "example": "uploads/audio/sample.opus"
                },
                "processing_time_ms": {
                    "type": "integer",
                    "example": 1200
                },
                "provider": {
                    "type": "string",
                    "example": "minio"
                },
                "size": {
                    "type": "integer",
                    "example": 7340032
                },
                "upload_id": {
                    "type": "string",
                    "example": "44c62b0d-7d55-4c74-9f65-8c7ab1f06642"
                },
                "url": {
                    "type": "string",
                    "example": "https://cdn.example.com/uploads/audio/sample.opus"
                },
                "version_id": {
                    "type": "string",
                    "example": "3/L4kqtJlcpXroDTDmJ+rmSpXd3dIbrHY"
                }
            }
        },
        "whats-convert-api_internal_models.S3UploadStatusResponse": {
            "type": "object",
            "properties": {
                "bytes_transferred": {
                    "type": "integer",
                    "example": 1048576
                },
                "end_time": {
                    "type": "string",
                    "example": "2024-03-31T12:00:10Z"
                },
                "error": {
                    "type": "string",
                    "example": "connection reset by peer"
                },
                "progress": {
                    "type": "number",
                    "example": 42.5
                },
                "result": {
                    "$ref": "#/definitions/whats-convert-api_internal_models.S3UploadResult"
                },
                "start_time": {
                    "type": "string",
                    "example": "2024-03-31T12:00:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "uploading"
                },
                "total_bytes": {
                    "type": "integer",
                    "example": 5242880
                },
                "upload_id": {
                    "type": "string",
                    "example": "3f99d60f-bd8d-49e6-9ecf-2fbc9e4adffe"
                }
            }
        },
...
        "whats-convert-api_internal_providers.ObjectInfo": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "etag": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "last_modified": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "storage_class": {
                    "type": "string"
                },
                "version_id": {
                    "type": "string"
                }
            }
        },
...
    }
}